+++ src/external/libhydrogen/impl/pwhash.h
@@ -70,10 +70,10 @@ hydro_pwhash_keygen(uint8_t master_key[hydro_pwhash_MASTERKEYBYTES])
 }
 
 int
-hydro_pwhash_deterministic(uint8_t *h, size_t h_len, const char *passwd, size_t passwd_len,
+hydro_pwhash_deterministic(uint8_t *h, size_t h_len, const char *passwd, int passwd_len,
                            const char    ctx[hydro_pwhash_CONTEXTBYTES],
-                           const uint8_t master_key[hydro_pwhash_MASTERKEYBYTES], uint64_t opslimit,
-                           size_t memlimit, uint8_t threads)
+                           const uint8_t master_key[hydro_pwhash_MASTERKEYBYTES], int opslimit,
+                           int memlimit, uint8_t threads)
 {
     uint8_t seed[hydro_random_SEEDBYTES];
